requirements.txt
anyio
bcrypt
fastapi
fastapi-cli
jose
jwt
passlib
pydantic
pydantic-settings
pydantic_core
PyJWT
pytest
pytest-asyncio
python-decouple
python-dotenv
python-jose
PyYAML
SQLAlchemy
uvicorn